backend/requirements.txt
openai>=1.54.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pymongo>=4.10.0
